add: distribute fees buyback script
